+++ CHANGELOG.md
@@ -41,6 +41,8 @@
   - [#1176](https://github.com/wix-incubator/rich-content/pull/1176) isMobile automatic determination
 - `ricos`
   - [#1172](https://github.com/wix-incubator/rich-content/pull/1172) ThemeStrategy throws error when Palette is given with missing colors
+- `test-env`
+  - [#1183](https://github.com/wix-incubator/rich-content/pull/1183) ricos coverage of file-upload + rename from wrapper
 
 </details>
 <hr/>